+++ victreebot/utils/helpers/contants.py
@@ -120,5 +120,5 @@ SUPPORTED_TIMEZONES = [
     "GMT+12",
 ]
 SUPPORTED_UNIT_SYSTEMS = ["Metric System", "Imperial System", "United States Customary Unit System"]
-SUPPORTED_RAID_TYPES = ["Raid", "Mega-Raid"]
+SUPPORTED_RAID_TYPES = ["Raid", "Mega-Raid", "Elite-Raid", "EX-Raid"]
 SUPPORTED_LOCATION_TYPES = ["Gym", "Pokéstop"]